% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-lib.R
\docType{data}
\name{AntaresClientConfigurator.class}
\alias{AntaresClientConfigurator.class}
\title{The class that holds all the configuration.
It is not accessible directly but through get_config_value}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AntaresClientConfigurator.class
}
\description{
The class that holds all the configuration.
It is not accessible directly but through get_config_value
}
\keyword{datasets}
