% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{AntaresRClient}
\alias{AntaresRClient}
\alias{.onLoad}
\alias{AntaresRClient-package}
\title{AntaresRClient}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
A client to interact with the Antares Infrastructure
}
\author{
Alejandro Baranek <abaranek@dc.uba.ar>, Leonardo Javier Belen <leobelen@gmail.com>
Executes code while loading the package.
}
