% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-lib.R
\name{get_config_value}
\alias{get_config_value}
\title{Gets a value from the configurator object or returns NA otherwise.}
\usage{
get_config_value(section, code)
}
\arguments{
\item{code}{Parameter code}

\item{default.value}{a default value in case it cannot be found.}

\item{is.dir}{triggers directories creation}
}
\description{
Gets a value from the configurator object or returns NA otherwise.
}
