% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antares-lib.R
\name{get_message_sender}
\alias{get_message_sender}
\title{Constructs an AntaresMessageSender.class}
\usage{
get_message_sender(url = NA, token = NA)
}
\description{
Constructs an AntaresMessageSender.class
}
